Chức năng 2 from: NV, PB

+++ QuanLyNganHang/ET/ET_NhanVien.cs
@@ -11,13 +11,13 @@ namespace ET
         private string _GioiTinh;
         private DateTime _NgaySinh;
         private string _Chuc;
-        private int _Luong;
+        private float _Luong;
         private string _DiaChi;
         private int _Sdt;
         private int _MaPB;
         private string _MaCN;
 
-        public ET_NhanVien(string maNV, string tenNV, string gioiTinh, DateTime ngaySinh, string chuc, int luong, string diaChi, int sdt, int maPB, string maCN)
+        public ET_NhanVien(string maNV, string tenNV, string gioiTinh, DateTime ngaySinh, string chuc, float luong, string diaChi, int sdt, int maPB, string maCN)
         {
             _MaNV = maNV;
             _TenNV = tenNV;
@@ -36,7 +36,7 @@ namespace ET
         public string GioiTinh { get => _GioiTinh; set => _GioiTinh = value; }
         public DateTime NgaySinh { get => _NgaySinh; set => _NgaySinh = value; }
         public string Chuc { get => _Chuc; set => _Chuc = value; }
-        public int Luong { get => _Luong; set => _Luong = value; }
+        public float Luong { get => _Luong; set => _Luong = value; }
         public string DiaChi { get => _DiaChi; set => _DiaChi = value; }
         public int Sdt { get => _Sdt; set => _Sdt = value; }
         public int MaPB { get => _MaPB; set => _MaPB = value; }
